+++ pertemuan_5dan6/lib/main.dart
@@ -1,46 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:pertemuan_5dan6/routes.dart';
 
 void main() {
   runApp(MaterialApp(
-    home: HomePage(),
+    onGenerateRoute: RouteGenerator.generateRoute,
   ));
-}
-
-class HomePage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Belajar Routing'),
-      ),
-      body: Center(
-        child: RaisedButton(
-          onPressed: () {
-            Route route = MaterialPageRoute(builder: (context) => AboutPage());
-            Navigator.push(context, route);
-          },
-          child: Text('Tap Untuk ke AboutPage'),
-        ),
-      ),
-    );
-  }
-}
-
-class AboutPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Tentang Aplikasi'),
-      ),
-      body: Center(
-        child: RaisedButton(
-          onPressed: () {
-            Navigator.pop(context);
-          },
-          child: Text('Kembali'),
-        ),
-      ),
-    );
-  }
 }
