+++ pertemuan_5dan6/lib/routes.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+import 'package:pertemuan_5dan6/screen.dart';
+
+class RouteGenerator {
+  static Route<dynamic> generateRoute(RouteSettings settings) {
+    // jika ingin mengirim argument
+    // final args = settings.arguments;
+    switch (settings.name) {
+      case '/':
+        return MaterialPageRoute(builder: (_) => HomePage());
+      case '/about':
+        return MaterialPageRoute(builder: (_) => AboutPage());
+      // return MaterialPageRoute(builder: (_) => AboutPage(args));
+      default:
+        return _errorRoute();
+    }
+  }
+
+  static Route<dynamic> _errorRoute() {
+    return MaterialPageRoute(builder: (_) {
+      return Scaffold(
+        appBar: AppBar(title: Text("Error")),
+        body: Center(child: Text('Error page')),
+      );
+    });
+  }
+}
